cleanup

+++ build/osx-dmginstaller/buildinstaller.sh
@@ -2,11 +2,11 @@
 ## this is shamelessly taken from Pd-extended...
 
 SRCDIR=$1
-VOLUME_NAME=Gem-0.93.3
-PACKAGE_NAME="${VOLUME_NAME}"
-DESTDIR="/Volumes/${VOLUME_NAME}"
+PACKAGE_NAME=${SRCDIR%/}
+VOLUME_NAME="${PACKAGE_NAME}"
+VOLUME_NAME="Gem Installer"
 
-VOLUME_NAME=${SRCDIR%/}
+DESTDIR="/Volumes/${VOLUME_NAME}"
 
 if [ -d "${SRCDIR}" ]; then
  echo "creating Gem installer from ${SRCDIR}"
@@ -33,22 +33,17 @@ USERINSTALLER="${DESTDIR}/Gem User-specific Installer.app"
 ./icon2app "${USERINSTALLER}" /System/Library/CoreServices/CoreTypes.bundle/Contents/Resources/UsersFolderIcon.icns
 
 # a manually crafted .DS_Store file
-install -p DS_Store "${DESTDIR}/.DS_Store"
+test  -e DS_Store && install -p DS_Store "${DESTDIR}/.DS_Store"
 
 # To enable the local image icon
 #install -p VolumeIcon.icns "${DESTDIR}/.VolumeIcon.icns"
 #/Developer/Tools/SetFile -a C "${DESTDIR}/.VolumeIcon.icns" "${DESTDIR}"
 
-#chmod -R a-w "${DESTDIR}/${PD_APP_NAME}.app/Contents/Resources"
-
+chmod -R a-w "${DESTDIR}/${PD_APP_NAME}.app/Contents/Resources"
 
 hdiutil detach $(mount | grep "${VOLUME_NAME}" | cut -d ' ' -f 1)
-hdiutil convert -format UDZO -o "${PACKAGE_NAME}.dmg" build.dmg
-
-exit 0
-
-rm -f build.dmg
-
+echo "now creating RO-image ${PACKAGE_NAME}.dmg from build.dmg"
+hdiutil convert -format UDZO -o "${PACKAGE_NAME}.dmg" build.dmg && rm -f build.dmg
 
 # install license
 hdiutil unflatten "${PACKAGE_NAME}.dmg"
